// FrontEnd/app/components/apartment/model/apartment.ts
export interface IApartment{
    id?: number;
    buildingName?: string;
    buildingBlock?: string;
    transactionType?: string;
    price?:number;
    propertyType?: string;
    area?: string;
    floorAndDirection?: string;
    buildingAge?: string;
    extraInfo1?: string;
    extraInfo2?: string;
    extraInfo3?: string;
    extraInfo4?: string;
    confirmationStatus?: string;
    date?: string;
    depositPrice?: number;
    monthPrice?:number;
}